extracted SaveButton & better visual feedback
