+++ Frontend/components/workspace.tsx
@@ -7,19 +7,13 @@ import ReactFlow, {
   MiniMap,
   Panel,
 } from "reactflow";
-import {
-  ArrowPathIcon,
-  CloudArrowDownIcon,
-  CloudIcon,
-  ExclamationTriangleIcon,
-  TrashIcon,
-} from "@heroicons/react/24/outline";
-import DistributionNode from "../components/DistributionNode";
+import DistributionNode from "../components/distributionNode";
 import { useStore, selector } from "../hooks/store";
 import CustomEdge from "../components/customEdge";
 import connectionLine from "./connectionLine";
 import "reactflow/dist/base.css";
-import { useUpdateModel } from "../hooks/useUpdateModel";
+import { TrashIcon } from "@heroicons/react/24/outline";
+import SaveButton from "./SaveButton";
 
 interface WorkspaceProps {
   className: string;
@@ -41,7 +35,6 @@ export default function Workspace({ className }: WorkspaceProps) {
   const [reactFlowInstance, setReactFlowInstance] = useState<any>(null);
   const { nodes, edges, onNodesChange, onEdgesChange, onConnect, addNode } =
     useStore(selector, shallow);
-  const updateModelMutation = useUpdateModel();
 
   const onDragOver = useCallback((event: any) => {
     event.preventDefault();
@@ -83,13 +76,6 @@ export default function Workspace({ className }: WorkspaceProps) {
     [reactFlowInstance]
   );
 
-  const handleSave = useCallback(() => {
-    if (reactFlowInstance) {
-      const flow = reactFlowInstance.toObject();
-      updateModelMutation.mutate({ id: 1, title: "test", body: flow });
-    }
-  }, [reactFlowInstance]);
-
   return (
     <div className={className} ref={reactFlowWrapper}>
       <ReactFlowProvider>
@@ -113,28 +99,7 @@ export default function Workspace({ className }: WorkspaceProps) {
             <button className="p-1 hover:bg-gray-200 rounded">
               <TrashIcon className="w-5" />
             </button>
-            <button
-              className="p-1 hover:bg-gray-200 rounded"
-              onClick={handleSave}
-            >
-              {updateModelMutation.isLoading ? (
-                <ArrowPathIcon className="w-5" />
-              ) : (
-                <>
-                  {updateModelMutation.isError ? (
-                    <ExclamationTriangleIcon className="w-5 text-red-500" />
-                  ) : (
-                    <>
-                      {updateModelMutation.isSuccess ? (
-                        <CloudIcon className="w-5 text-green-500" />
-                      ) : (
-                        <CloudArrowDownIcon className="w-5" />
-                      )}
-                    </>
-                  )}
-                </>
-              )}
-            </button>
+            <SaveButton reactFlowInstance={reactFlowInstance} />
           </Panel>
           <Background />
         </ReactFlow>
