+++ Frontend/hooks/useUpdateModel/index.ts
@@ -2,7 +2,6 @@ import { useMutation, useQuery } from "@tanstack/react-query";
 import ky from "ky-universal";
 
 type updatePayload = {
-  id: number;
   title: string;
   body: object;
 };
@@ -10,11 +9,12 @@ type updatePayload = {
 const useUpdateModel = () => {
   return useMutation({
     mutationFn: (payload: updatePayload) => {
-      return ky
-        .post(`${process.env.NEXT_PUBLIC_API_URL}/models/`, {
-          json: payload,
-        })
-        .json();
+      return ky.put("http://httpbin.org/anything", { json: payload }).json();
+      // return ky
+      //   .post(`${process.env.NEXT_PUBLIC_API_URL}/models/`, {
+      //     json: payload,
+      //   })
+      //   .json();
     },
   });
 };
